# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export type Severity = typeof Severity[keyof typeof Severity];


export const Severity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  safe: 'safe',
} as const;

export type RiskType = typeof RiskType[keyof typeof RiskType];


export const RiskType = {
  direct_prompt_injection: 'direct_prompt_injection',
  indirect_injection: 'indirect_injection',
  data_exfiltration: 'data_exfiltration',
  unsafe_tool_use: 'unsafe_tool_use',
  memory_poisoning: 'memory_poisoning',
  prompt_extraction: 'prompt_extraction',
} as const;

export type CaseStatus = typeof CaseStatus[keyof typeof CaseStatus];


export const CaseStatus = {
  new: 'new',
  investigating: 'investigating',
  contained: 'contained',
  closed: 'closed',
} as const;

export interface TraceInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 20000
     */
  content: string;
  /** @maxLength 80 */
  source?: string;
}

export interface CaseUpdate {
  status?: CaseStatus;
  /** @maxLength 4000 */
  analystNotes?: string;
}

export interface Finding {
  id: string;
  riskType: RiskType;
  label: string;
  severity: Severity;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  evidence: string;
  start: number;
  end: number;
  owasp: string;
  impact: string;
  mitigation: string;
}

export interface TimelineEvent {
  time: string;
  stage: string;
  detail: string;
  outcome: string;
}

export interface IncidentCase {
  id: string;
  title: string;
  source: string;
  content: string;
  createdAt: string;
  updatedAt: string;
  status: CaseStatus;
  analystNotes: string;
  severity: Severity;
  score: number;
  findings: Finding[];
  timeline: TimelineEvent[];
}

export interface CaseSummary {
  id: string;
  title: string;
  source: string;
  createdAt: string;
  status: CaseStatus;
  severity: Severity;
  score: number;
  findingCount: number;
}

export interface DemoTrace {
  id: string;
  label: string;
  /** @nullable */
  riskType: string | null;
  description: string;
  content: string;
  expectedMalicious: boolean;
}

export interface RiskCount {
  riskType: RiskType;
  label: string;
  count: number;
}

export interface Dashboard {
  totalCases: number;
  openCases: number;
  highRiskCases: number;
  averageScore: number;
  riskCounts: RiskCount[];
  recentCases: CaseSummary[];
}

export interface EvaluationRow {
  riskType: RiskType;
  label: string;
  truePositive: number;
  falsePositive: number;
  falseNegative: number;
  precision: number;
  recall: number;
}

export interface Evaluation {
  datasetSize: number;
  precision: number;
  recall: number;
  falsePositives: number;
  rows: EvaluationRow[];
  methodology: string;
}

export type ListCasesParams = {
search?: string;
severity?: Severity;
};