# lib/api-zod/src/generated/types/caseSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */
import type { CaseStatus } from './caseStatus';
import type { Severity } from './severity';

export interface CaseSummary {
  id: string;
  title: string;
  source: string;
  createdAt: Date;
  status: CaseStatus;
  severity: Severity;
  score: number;
  findingCount: number;
}