# lib/api-zod/src/generated/types/caseUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */
import type { CaseStatus } from './caseStatus';

export interface CaseUpdate {
  status?: CaseStatus;
  /** @maxLength 4000 */
  analystNotes?: string;
}