# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */
import type { CaseSummary } from './caseSummary';
import type { RiskCount } from './riskCount';

export interface Dashboard {
  totalCases: number;
  openCases: number;
  highRiskCases: number;
  averageScore: number;
  riskCounts: RiskCount[];
  recentCases: CaseSummary[];
}