# lib/api-zod/src/generated/types/demoTrace.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */

export interface DemoTrace {
  id: string;
  label: string;
  /** @nullable */
  riskType: string | null;
  description: string;
  content: string;
  expectedMalicious: boolean;
}