# lib/api-zod/src/generated/types/evaluation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */
import type { EvaluationRow } from './evaluationRow';

export interface Evaluation {
  datasetSize: number;
  precision: number;
  recall: number;
  falsePositives: number;
  rows: EvaluationRow[];
  methodology: string;
}