# lib/api-zod/src/generated/types/evaluationRow.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */
import type { RiskType } from './riskType';

export interface EvaluationRow {
  riskType: RiskType;
  label: string;
  truePositive: number;
  falsePositive: number;
  falseNegative: number;
  precision: number;
  recall: number;
}