# lib/api-zod/src/generated/types/finding.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */
import type { RiskType } from './riskType';
import type { Severity } from './severity';

export interface Finding {
  id: string;
  riskType: RiskType;
  label: string;
  severity: Severity;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  evidence: string;
  start: number;
  end: number;
  owasp: string;
  impact: string;
  mitigation: string;
}