# lib/api-zod/src/generated/types/incidentCase.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */
import type { CaseStatus } from './caseStatus';
import type { Finding } from './finding';
import type { Severity } from './severity';
import type { TimelineEvent } from './timelineEvent';

export interface IncidentCase {
  id: string;
  title: string;
  source: string;
  content: string;
  createdAt: Date;
  updatedAt: Date;
  status: CaseStatus;
  analystNotes: string;
  severity: Severity;
  score: number;
  findings: Finding[];
  timeline: TimelineEvent[];
}