# lib/api-zod/src/generated/types/riskCount.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */
import type { RiskType } from './riskType';

export interface RiskCount {
  riskType: RiskType;
  label: string;
  count: number;
}