# lib/api-zod/src/generated/types/riskType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */

export type RiskType = typeof RiskType[keyof typeof RiskType];


export const RiskType = {
  direct_prompt_injection: 'direct_prompt_injection',
  indirect_injection: 'indirect_injection',
  data_exfiltration: 'data_exfiltration',
  unsafe_tool_use: 'unsafe_tool_use',
  memory_poisoning: 'memory_poisoning',
  prompt_extraction: 'prompt_extraction',
} as const;