# lib/api-zod/src/generated/types/severity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */

export type Severity = typeof Severity[keyof typeof Severity];


export const Severity = {
  critical: 'critical',
  high: 'high',
  medium: 'medium',
  low: 'low',
  safe: 'safe',
} as const;