# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */

export interface TimelineEvent {
  time: string;
  stage: string;
  detail: string;
  outcome: string;
}