# lib/api-zod/src/generated/types/traceInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * AgentTripwire local AI-agent trace analysis API
 * OpenAPI spec version: 1.0.0
 */

export interface TraceInput {
  /**
     * @minLength 1
     * @maxLength 120
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 20000
     */
  content: string;
  /** @maxLength 80 */
  source?: string;
}